Former-commit-id: 5e9fa8e21f9efea8c5b5147c6c331c73ed36c625

+++ deeplearning4j-core/src/main/java/org/deeplearning4j/autoencoder/AutoEncoderOptimizer.java
@@ -9,7 +9,7 @@ import org.deeplearning4j.optimize.NeuralNetworkOptimizer;
  * @author Adam Gibson
  */
 public class AutoEncoderOptimizer extends NeuralNetworkOptimizer {
-    public AutoEncoderOptimizer(NeuralNetwork network, double lr, Object[] trainingParams, NeuralNetwork.OptimizationAlgorithm optimizationAlgorithm, NeuralNetwork.LossFunction lossFunction) {
+    public AutoEncoderOptimizer(NeuralNetwork network, float lr, Object[] trainingParams, NeuralNetwork.OptimizationAlgorithm optimizationAlgorithm, NeuralNetwork.LossFunction lossFunction) {
         super(network, lr, trainingParams, optimizationAlgorithm, lossFunction);
     }
