+++ linear-algebra/linear-algebra-api/src/main/java/org/deeplearning4j/linalg/ops/elementwise/MultiplyOp.java
@@ -36,7 +36,7 @@ public class MultiplyOp extends BaseTwoArrayElementWiseOp {
 
             //real + complex
             else {
-                double element = (double) origin.element();
+                float element = (float) origin.element();
                 IComplexNumber result = otherValue.mul(element);
                 return NDArrays.scalar(result);
 
@@ -50,7 +50,7 @@ public class MultiplyOp extends BaseTwoArrayElementWiseOp {
             if(origin instanceof IComplexNDArray) {
                 IComplexNDArray originComplexValue = (IComplexNDArray) origin;
                 IComplexNumber firstValue = (IComplexNumber) originComplexValue.element();
-                double realValue = (double) value.element();
+                float realValue = (float) value.element();
                 IComplexNumber retValue = firstValue.mul(realValue);
                 return NDArrays.scalar(retValue);
 
@@ -58,8 +58,8 @@ public class MultiplyOp extends BaseTwoArrayElementWiseOp {
 
             //both normal
             else {
-                double firstElement = (double) origin.element();
-                double secondElement = (double) value.element();
+                float firstElement = (float) origin.element();
+                float secondElement = (float) value.element();
                 return NDArrays.scalar(firstElement * secondElement,0);
             }
