+++ linear-algebra/linear-algebra-api/src/main/java/org/deeplearning4j/linalg/ops/reduceops/scalarops/ScalarOp.java
@@ -7,5 +7,5 @@ import org.deeplearning4j.linalg.api.ndarray.INDArray;
  * A scalar op over an ndarray (iterates through the whole ndarray for an aggregate result)
  * @author Adam Gibson
  */
-public interface ScalarOp extends Function<INDArray,Double> {
+public interface ScalarOp extends Function<INDArray,Float> {
 }
